♻️ Refactor: carousel 컴포넌트 모듈로 관리되게 변경
refactor: carousel 컴포넌트 모듈로 관리되게 변경
