+++ src/shared/ui/index.ts
@@ -14,3 +14,4 @@ export * from "./timeSelectBox";
 export * from "./dateSelectBox";
 export * from "./shadcnInput";
 export * from "./textarea";
+export * from "./carousel";
